Revert "Improve documentation of `disable_init` decorator"
The new recommendation actually doesn't work. The correct way to do it
would be:

```py
super(cls, self).__init__()
```

But this is not even necessary when you are applying the decorator to
a simple class that doesn't inherit from anything (i.e. inherits from
`object` directly).

Applying the decorators to other sub-classes, is much trickier than just
addint the `super()` call, so we will just not recommend or support
doing that and keep the usage documentation simple.

This reverts commit 073b9216ee2c7c1051494168a01298b14a32dc7b.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>