typing: Improve documentation
Improve the docs to point out that disabling the `__init__` for existing
classes via a sub-class is complicated, so it is only recommended to use
both the decorator and metaclass on classes inheriting from `object`
directly.

Also improve the documentation of `NoInitConstructibleMeta` to be more
complete and give an example on using with `abc.ABCMeta`.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>