Generalize `logging.get_public_logger` to be about module names
This module and function were too specific to `logging`, when it could
be more generally useful if it just deal with module names instead.

The module was renamed to `core.module` and the function to
`get_public_module_name()`, and it now deals only with strings.

When this is needed to get a logger, users can simply do:

logger = logging.getLogger(module.get_public_module_name(__name__))

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>